iot_alarm_server/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
psycopg[binary]==3.2.6
paho-mqtt==2.1.0
jinja2==3.1.6
